requirement.txt
torch==2.2.0
torchview==0.2.6
onnx==1.15.0
numpy==1.24.4
transformers==4.37.2
pandas==2.0.3
graphviz==0.20.1
pillow==10.2.0
timm==0.9.12
scikit-learn==1.3.2
loguru==0.7.2
tqdm==4.66.1
python-dotenv==1.0.1
matplotlib==3.7.4
openai==1.12.0
bitsandbytes==0.42.0
accelerate==0.27.2
openpyxl==3.1.2